requirements.txt
streamlit
openai
streamlit-mic-recorder
edge-tts
Pillow
